Apply top nav with entire page semantic structure
The aside tag is added as this page is link to existing design system stylesheets which uses grid for users to navigate mainly within the main content. So aside is set to empty as not required in the Dev Page. This works for the main content to go to main grid box